Almost done formatting, just -po is wrong (see commit desc)
-po is wrong because it is counting frequency across both files. Ex. if
there is 1 hello in simpleTest and 1 in otherTest, it will show as "2
found in file: simpleTest on lines 1" and the same for the otherTest
line. Speaking of the fileNames being on two different lines, I need to
get the filenames showing up on the same line. I just am too tired to do
that right now. Good thing it's only wednesday. Whew!